move command logic to app project
