+++ src/Microsoft.DocAsCode.App/Helpers/DocumentBuilderWrapper.cs
