+++ src/Microsoft.DocAsCode.App/RunMerge.cs
@@ -1,0 +1,106 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.IO;
+using Microsoft.DocAsCode.Build.Engine;
+using System.Linq;
+using Microsoft.DocAsCode.Common;
+using Microsoft.DocAsCode.Plugins;
+using Microsoft.DocAsCode.SubCommands;
+
+namespace Microsoft.DocAsCode
+{
+    internal static class RunMerge
+    {
+        public static void Exec(MergeJsonConfig config)
+        {
+            foreach (var round in config)
+            {
+                var baseDirectory = round.BaseDirectory ?? Directory.GetCurrentDirectory();
+                var intermediateOutputFolder = round.Destination ?? Path.Combine(baseDirectory, "obj");
+                EnvironmentContext.SetBaseDirectory(baseDirectory);
+                EnvironmentContext.SetOutputDirectory(intermediateOutputFolder);
+                MergeDocument(config, baseDirectory, intermediateOutputFolder);
+                EnvironmentContext.Clean();
+            }
+        }
+
+        private static void MergeDocument(MergeJsonConfig config, string baseDirectory, string outputDirectory)
+        {
+            foreach (var round in config)
+            {
+                var parameters = ConfigToParameter(round, baseDirectory, outputDirectory);
+                if (parameters.Files.Count == 0)
+                {
+                    Logger.LogWarning("No files found, nothing is to be generated");
+                    continue;
+                }
+                try
+                {
+                    new MetadataMerger().Merge(parameters);
+                }
+                catch (Exception ex)
+                {
+                    Logger.LogError(ex.ToString());
+                }
+            }
+        }
+
+        private static MetadataMergeParameters ConfigToParameter(MergeJsonItemConfig config, string baseDirectory, string outputDirectory) =>
+            new MetadataMergeParameters
+            {
+                OutputBaseDir = outputDirectory,
+                Metadata = config.GlobalMetadata?.ToImmutableDictionary() ?? ImmutableDictionary<string, object>.Empty,
+                FileMetadata = ConvertToFileMetadataItem(baseDirectory, config.FileMetadata),
+                TocMetadata = config.TocMetadata?.ToImmutableList() ?? ImmutableList<string>.Empty,
+                Files = GetFileCollectionFromFileMapping(
+                    baseDirectory,
+                    DocumentType.Article,
+                    GlobUtility.ExpandFileMapping(baseDirectory, config.Content)),
+            };
+
+        private static FileMetadata ConvertToFileMetadataItem(string baseDirectory, Dictionary<string, FileMetadataPairs> fileMetadata)
+        {
+            if (fileMetadata == null)
+            {
+                return null;
+            }
+            var result = new Dictionary<string, ImmutableArray<FileMetadataItem>>();
+            foreach (var item in fileMetadata)
+            {
+                var list = new List<FileMetadataItem>();
+                foreach (var pair in item.Value.Items)
+                {
+                    list.Add(new FileMetadataItem(pair.Glob, item.Key, pair.Value));
+                }
+                result.Add(item.Key, list.ToImmutableArray());
+            }
+
+            return new FileMetadata(baseDirectory, result);
+        }
+
+        private static IEnumerable<string> GetFilesFromFileMapping(FileMapping mapping)
+        {
+            if (mapping == null)
+            {
+                return Enumerable.Empty<string>();
+            }
+            return from file in mapping.Items
+                   from item in file.Files
+                   select Path.Combine(file.SourceFolder ?? Directory.GetCurrentDirectory(), item);
+        }
+
+        private static FileCollection GetFileCollectionFromFileMapping(string baseDirectory, DocumentType type, FileMapping files)
+        {
+            var result = new FileCollection(baseDirectory);
+            foreach (var mapping in files.Items)
+            {
+                result.Add(type, mapping.Files, mapping.SourceFolder, mapping.DestinationFolder);
+            }
+            return result;
+        }
+    }
+}
